Refactor consent check to use CmpErrors for feedback
Standard consent record submissions now give feedback about where the
error was. The refactor changes the validation functions so that they
return either the value or an error describing why the value was
wrong.

At the same time, we also now properly validate the types of the
provided data. Up to now we'd been assuming that the provided JSON
structure had the correct types in some cases.

The feedback still needs improving for the purposes sub-object, but
this is already a huge improvement in usability.