Tighten up CMP checks in tests, and fix bug
We now extend jest to provide some useful helpers for checking whether
results are CmpErrors. This test combines the previous checks with an
assertion that the error message is not empty.

This more detailed test caught a bug in the implementation that was
leading to blank error messages, that bug is also fixed here in the
`src/errors.ts` file.