Refactor responses so they are always returned to the user
The first parameter of the callback if to signal to AWS that the
execution was a complete failure, and this instructs API Gateway to
return a blank 502. This isn't what we want! We actually want to
return the JSON, even in "error" states. To accomodate this behaviour,
the responses have been refactored to use the callbacks, with log
lines to give more information about the failures.