A bunch of small tidy-ups
Switch to node 10.x (from 8.10). The tests only ran on node 10 but the
Lambda was using the node8 runtime in PROD. This unifies the AWS and
dev environments on node 10.

Removes code that was left over from the time when there was a local
dev server. We may restore that functionality, but for now it isn't
needed.

The TeamCity build now also runs the tests.