Further improvements, removes time from submission
This further tidies the codebase, including fixing som unimplemented
tests. The main change is the removal of the time field from
submissions. This meant trusting the time given by the client, which
we don't want to do in this case. The time is now added by the server
as part of processing the submission record.