Refactor to allow versioned validation
The model has been separated from the validation functions for
clarity. The models have been re-worked so they are (fairly) consisely
defined in one place at the type and value level. Multiple versions of
the model are available to cater for records that do not contain the
PECR purposes.

Validation has been split out into its own module. There are now
multiple purpose validation functions to deal with the two known
versions (with and without PECR purposes). There are overall
validation functions for both versions that delegate to tthe correct
field validation, and a typescript exhaustiveness check is in place to
make sure all versions are handled.

The validation tests have been tidied. The parseJson tests are now
concerned with the overall success rather than the details for each
key. The tests that look in detail at each key are mostly moved to
tests for each of the focused validation functions.