Update initial sources for the test rollout
While working on the docs I spotted the sources aren't very accurate
at the moment. Rather than speculate about what should go here, this
sets it to `www` and `amp`, which are the two concrete requirements
for now.

Additional subdomains and the live apps can be added as they are
required.