Introduces CORS support to the consent logs
Adds CORS headers to the responses, as well as origin checks to the
handler. The origin checks are informed by a new piece of
configuration, drawn from the CloudFormation (see previous commit). We
add support for checking the origin header against the configured
whitelist. If it matches then the appropriate CORS headers are added
to the response. If not, an HTTP 403 (forbidden) is returned.

The change is made up of three main parts:
* CORS whitelist obtained from ENV (via Cfn)
* CORS checks wired into handler and response helpers
* cors module that contains logic and helpers for the CORS functionality

The latter includes tests in `cors.test.spec`.